first cgommit

+++ src/dwarf/graphics/util.java
@@ -15,11 +15,30 @@ public final class util {
                 "you can not instantiate this class.");
     }
 
-    public static void setBackgroundColour(int red, int green, int blue) {
+    /**
+     * Set the background colour of the screen. The red,green,blue colour
+     * components are byte in the range 0-255. Black is 0,0,0 and white is
+     * 255,255,255. Colour will take effect the next time the screen is cleared.
+     *
+     * @param red red component (byte)
+     * @param green green component (byte)
+     * @param blue blue component (byte)
+     */
+    public static void setBackgroundColour(byte red, byte green, byte blue) {
         util.backgroundColour = new Colour(red, green, blue);
     }
 
-    public static void setBackgroundColour(int red, int green, int blue, int alpha) {
+    /**
+     * Set the background colour of the screen. The red,green,blue colour
+     * components are byte in the range 0-255. Black is 0,0,0 and white is
+     * 255,255,255. Colour will take effect the next time the screen is cleared.
+     *
+     * @param red red component (byte)
+     * @param green green component (byte)
+     * @param blue blue component (byte)
+     * @param alpha alpha component (byte)
+     */
+    public static void setBackgroundColour(byte red, byte green, byte blue, byte alpha) {
 
         util.backgroundColour = new Colour(red, green, blue, alpha);
     }
